requirements.txt
streamlit
pandas
requests
python-dotenv
openai
zhipuai
aiohttp
email-validator
dnspython
beautifulsoup4
yagmail
jinja2
httpx
schedule
matplotlib
